Migrate to create-react-app
Was hoping to try to go for a setup that didn't depend on CRA, but
parceljs had issues with producing a production build size that was
acceptable, and Rollup seemed too complex to work out.

I guess that at the end of the day if I was really going to get react
working from scratch I would go with Rollup and figure it all out, but
for now I am happy to hide all the convoluted dependencies and babel
transformations behind the CRA abstraction and hope it doesn't come
back to bite me.